+++ android/app/src/main/kotlin/com/chulminator/jubu/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.jubu
+﻿package com.chulminator.jubu
 
 import io.flutter.embedding.android.FlutterActivity